refactor(landing): restructure landing page components and tabs
