--- apps/landing/src/blocks/features-block.astro
@@ -1,81 +1,0 @@
----
-const features = [
-   {
-      title: "Pague pelo uso",
-      desc: "Sem assinatura fixa. Você paga pelos eventos que custam — IA, e-mail, storage. UI e CRUD são grátis.",
-      icon: "M3 12h4l2-7 4 14 2-7h6",
-   },
-   {
-      title: "Construído no Brasil",
-      desc: "CPF, CNPJ, Pix, OFX, NFe própria. Pensado para a realidade fiscal e bancária local.",
-      icon: "M12 2L2 7v10l10 5 10-5V7L12 2zM12 22V12M12 12L2 7M12 12l10-5",
-   },
-   {
-      title: "Open by design",
-      desc: "API oRPC tipada, webhooks, MCP. Integre com seus agentes de IA ou pipelines existentes.",
-      icon: "M16 18l6-6-6-6M8 6l-6 6 6 6",
-   },
-   {
-      title: "Rubi nativa",
-      desc: "Não é chatbot grudado depois. A agente foi projetada junto com o ERP — ela enxerga tudo.",
-      icon: "M12 2a10 10 0 100 20 10 10 0 000-20zM8 14s1.5 2 4 2 4-2 4-2M9 9h.01M15 9h.01",
-   },
-   {
-      title: "Workflows duráveis",
-      desc: "DBOS por baixo. Automações que não perdem estado em deploy, falha ou retry.",
-      icon: "M21 12a9 9 0 11-18 0 9 9 0 0118 0zM12 7v5l3 3",
-   },
-   {
-      title: "Hospedado por você ou nós",
-      desc: "Cloud gerenciada ou self-hosted. O código é um monorepo — nada escondido.",
-      icon: "M5 12V7a2 2 0 012-2h10a2 2 0 012 2v5M3 12h18l-2 7H5l-2-7z",
-   },
-];
----
-
-<section
-   id="rubi"
-   class="mx-auto max-w-7xl px-4 py-24 sm:px-6"
-   aria-label="Por que Montte"
->
-   <div class="mx-auto mb-16 flex max-w-3xl flex-col items-center gap-4 text-center">
-      <span
-         class="inline-flex items-center gap-2 rounded-full border border-border/60 bg-card/60 px-3 py-1 text-xs font-medium text-muted-foreground"
-      >
-         Por que Montte
-      </span>
-      <h2
-         class="text-4xl font-semibold tracking-[-0.03em] text-foreground text-balance sm:text-5xl"
-      >
-         O ERP que parece ter sido feito em 2026.
-      </h2>
-   </div>
-
-   <div class="grid gap-px overflow-hidden rounded-3xl border border-border/60 bg-border/40 sm:grid-cols-2 lg:grid-cols-3">
-      {
-         features.map((f) => (
-            <div class="group relative bg-background/40 p-8 backdrop-blur transition-colors hover:bg-card/60">
-               <div class="mb-5 inline-flex size-10 items-center justify-center rounded-xl bg-primary/15 text-primary ring-1 ring-primary/30">
-                  <svg
-                     class="size-5"
-                     viewBox="0 0 24 24"
-                     fill="none"
-                     stroke="currentColor"
-                     stroke-width="2"
-                     stroke-linecap="round"
-                     stroke-linejoin="round"
-                  >
-                     <path d={f.icon} />
-                  </svg>
-               </div>
-               <h3 class="mb-2 text-lg font-semibold text-foreground">
-                  {f.title}
-               </h3>
-               <p class="text-sm leading-relaxed text-muted-foreground">
-                  {f.desc}
-               </p>
-            </div>
-         ))
-      }
-   </div>
-</section>
