--- apps/landing/src/blocks/footer-block.astro
@@ -1,76 +1,0 @@
----
-import { Image } from "astro:assets";
-
-const groups = [
-   {
-      title: "Produto",
-      links: [
-         { label: "Financeiro", href: "#produto" },
-         { label: "Contatos", href: "#produto" },
-         { label: "Serviços", href: "#produto" },
-         { label: "Rubi IA", href: "#rubi" },
-      ],
-   },
-   {
-      title: "Empresa",
-      links: [
-         { label: "Sobre", href: "#" },
-         { label: "Blog", href: "#" },
-         { label: "Contato", href: "#" },
-      ],
-   },
-   {
-      title: "Recursos",
-      links: [
-         { label: "Docs", href: "#" },
-         { label: "Status", href: "#" },
-         { label: "Mudanças", href: "#" },
-      ],
-   },
-];
----
-
-<footer class="mt-16 border-t border-border/40 bg-background/40 backdrop-blur">
-   <div class="mx-auto grid max-w-7xl gap-12 px-4 py-16 sm:px-6 lg:grid-cols-[2fr_3fr]">
-      <div class="flex flex-col gap-4">
-         <a class="inline-flex items-center gap-2" href="/" aria-label="Montte">
-            <span class="inline-flex size-7 items-center justify-center rounded-lg bg-primary/15 ring-1 ring-primary/30">
-                <Image class="size-4" src="/favicon.svg" alt="" width={16} height={16} />
-            </span>
-            <span class="text-base font-semibold tracking-tight">Montte</span>
-         </a>
-         <p class="max-w-sm text-sm leading-relaxed text-muted-foreground">
-            ERP brasileiro construído ao redor de uma agente de IA.
-         </p>
-      </div>
-
-      <div class="grid grid-cols-2 gap-8 sm:grid-cols-3">
-         {
-            groups.map((g) => (
-               <div class="flex flex-col gap-3">
-                  <span class="text-xs font-semibold tracking-wider text-foreground uppercase">
-                     {g.title}
-                  </span>
-                  <ul class="flex flex-col gap-2">
-                     {g.links.map((l) => (
-                        <li>
-                           <a
-                              href={l.href}
-                              class="text-sm text-muted-foreground transition-colors hover:text-foreground"
-                           >
-                              {l.label}
-                           </a>
-                        </li>
-                     ))}
-                  </ul>
-               </div>
-            ))
-         }
-      </div>
-   </div>
-   <div
-      class="border-t border-border/40 px-4 py-6 text-center text-xs text-muted-foreground sm:px-6"
-   >
-      © {new Date().getFullYear()} Montte. Feito no Brasil.
-   </div>
-</footer>
