--- apps/landing/src/blocks/hero-block.astro
@@ -1,55 +1,0 @@
----
-import { Image } from "astro:assets";
-import { Picture } from "astro:assets";
-import { LandingHeroActions } from "../components/landing-hero-actions";
-import { HeroPosthogTabs } from "../components/hero-posthog-tabs.tsx";
-import heroRight from "../../public/illustrations/hero-right.png";
----
-
-<section
-   class="relative mx-auto flex max-w-7xl flex-col gap-16 px-4 pt-16 pb-12 sm:px-6 sm:pt-24"
-   aria-label="Apresentação"
->
-   <div class="grid items-start gap-12 lg:grid-cols-[minmax(0,1fr)_520px]">
-      <div class="flex flex-col items-start gap-4">
-         <a
-            class="inline-flex items-center gap-4 text-3xl font-black tracking-[-0.04em] text-foreground"
-            href="/"
-            aria-label="Montte"
-         >
-            <Image class="size-10" src="/favicon.svg" alt="" width={40} height={40} />
-            Montte
-         </a>
-
-         <div class="flex flex-col gap-4 pt-6">
-            <h1
-               class="max-w-3xl text-4xl leading-tight font-black tracking-[-0.05em] text-foreground text-balance sm:text-5xl"
-            >
-               A nova forma de operar seu ERP
-            </h1>
-
-            <p class="max-w-3xl text-base leading-relaxed text-muted-foreground">
-               Operar uma empresa costumava significar alternar entre planilhas,
-               relatórios, cobranças, financeiro e dezenas de ferramentas.
-            </p>
-            <p class="max-w-3xl text-base leading-relaxed text-muted-foreground">
-               O Montte é a plataforma que atua como copiloto para você e seus
-               agentes de IA operarem seu ERP com autonomia.
-            </p>
-         </div>
-
-          <LandingHeroActions client:load />
-      </div>
-
-      <div class="relative hidden h-96 items-start justify-end lg:flex" aria-label="Ilustração do Montte">
-         <Picture
-            class="h-full w-auto object-contain"
-            src={heroRight}
-            formats={["avif", "webp"]}
-            alt="Ilustração da Rubi operando o Montte"
-         />
-      </div>
-   </div>
-
-    <HeroPosthogTabs client:load />
-</section>
