--- apps/landing/src/blocks/product-tabs-block.astro
@@ -1,28 +1,0 @@
----
-import { ProductTabs } from "../components/product-tabs.tsx";
----
-
-<section
-   id="produto"
-   class="relative mx-auto max-w-7xl px-4 py-24 sm:px-6"
-   aria-label="Produto"
->
-   <div class="mx-auto mb-12 flex max-w-3xl flex-col items-start gap-4">
-      <span
-         class="inline-flex items-center gap-2 rounded-full border border-border/60 bg-card/60 px-3 py-1 text-xs font-medium text-muted-foreground"
-      >
-         O produto
-      </span>
-      <h2
-         class="text-4xl font-semibold tracking-[-0.03em] text-foreground text-balance sm:text-5xl"
-      >
-         Construído na forma do seu negócio.
-      </h2>
-      <p class="max-w-2xl text-lg leading-relaxed text-muted-foreground">
-         De medidores customizados a workflows duráveis, o Montte se molda à
-         realidade de empresas brasileiras — e a Rubi opera o resto.
-      </p>
-   </div>
-
-   <ProductTabs client:visible />
-</section>
