--- apps/landing/src/blocks/theme-switcher.astro
@@ -1,73 +1,0 @@
----
-const themes = [
-   {
-      key: "system",
-      label: "Sistema",
-   },
-   {
-      key: "light",
-      label: "Claro",
-   },
-   {
-      key: "dark",
-      label: "Escuro",
-   },
-];
----
-
-<div class="relative isolate flex rounded-full bg-background p-2 ring-1 ring-border" data-theme-switcher>
-   {
-      themes.map((theme) => (
-         <button
-            aria-label={`Tema ${theme.label.toLowerCase()}`}
-            class="rounded-full px-2 py-2 text-xs font-medium text-muted-foreground transition-colors hover:bg-accent hover:text-accent-foreground data-[active=true]:bg-muted data-[active=true]:text-foreground"
-            data-theme-option={theme.key}
-            type="button"
-         >
-            {theme.label}
-         </button>
-      ))
-   }
-</div>
-
-<script is:inline>
-   (() => {
-      const switcher = document.querySelector("[data-theme-switcher]");
-      if (!switcher) return;
-
-      const options = switcher.querySelectorAll("[data-theme-option]");
-      const getStoredTheme = () => {
-         const storedTheme = localStorage.getItem("montte:theme");
-         if (storedTheme === "dark" || storedTheme === "light") return storedTheme;
-         return "system";
-      };
-      const applyTheme = (theme) => {
-         const resolvedTheme =
-            theme === "system"
-               ? window.matchMedia("(prefers-color-scheme: dark)").matches
-                  ? "dark"
-                  : "light"
-               : theme;
-
-         document.documentElement.classList.remove("light", "dark");
-         document.documentElement.classList.add(resolvedTheme);
-         localStorage.setItem("montte:theme", theme);
-         document.cookie = `theme=${theme}; path=/; max-age=31536000; SameSite=Lax`;
-
-         for (const option of options) {
-            option.setAttribute(
-               "data-active",
-               option.getAttribute("data-theme-option") === theme ? "true" : "false",
-            );
-         }
-      };
-
-      applyTheme(getStoredTheme());
-
-      for (const option of options) {
-         option.addEventListener("click", () => {
-            applyTheme(option.getAttribute("data-theme-option"));
-         });
-      }
-   })();
-</script>
